pyproject.toml
[project]
name = "web-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "itsdangerous>=2.2.0",
    "passlib>=1.7.4",
    "pydantic-extra-types>=2.10.5",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "python-jose>=3.5.0",
    "sqladmin>=0.21.0",
    "sqlalchemy>=2.0.43",
]
